Interceptors to preserve & restore trace headers
When a request is sent to a server running on Cloud Run, a span will be
created automatically by Cloud Run. However, it only gets sampled to be
uploaded about 1 QPS.

This breaks the span hierarchy. To avoid this, we save the GRPC client's
outgoing traceparent header with a different name and restore it on the
GRPC server side. Effectively we skip the Cloud Run generated span
entirely, but preserving the span hierarchy all the time.

Signed-off-by: Nghia Tran <tcnghia@gmail.com>